final edition
added some id for some tags. if a user just join in a group the page will jump to myGroupDetail(member).html, but if he creates a new group successfully, the page will jump to myGroupDetails(organiser).html.   myGroups is a new page which shows all the groups where the user is.